+++ src/person/daizhongde/migration/spring/service/impl/MigJobProcessServiceImpl.java
@@ -1130,7 +1130,7 @@ public class MigJobProcessServiceImpl implements MigJobProcessService {
 		
 		return (MigJobProcess)dataDAO.listAllByMap( "from MigJobProcess t1 where "+hqlA.getWhereBackHQL(), hqlA.getMap() ).get(0);
 	}
-	public MigJobProcess browsePOJOById(int id) {
+	public MigJobProcess browsePOJOById(long id) {
 //		return dataDAO.findById((short)id);
 		return null;
 	}
